+++ yaqc/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- Support for nested named types
+
 ## [2020.07.3]
 
 ### Changed