+++ yaqc/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [2020.10.0]
+
 ### Added
 - Support for nested named types
 - Clients will reconnect on failure, with callbacks for action needed upon reconnect
@@ -82,7 +84,8 @@ There are no actual code changes, this release is to update the release pipeline
 ### Added
 - initial release
 
-[Unreleased]: https://gitlab.com/yaq/yaq-python/-/compare/yaqc-2020.07.3...master
+[Unreleased]: https://gitlab.com/yaq/yaq-python/-/compare/yaqc-2020.10.0...master
+[2020.10.0]: https://gitlab.com/yaq/yaq-python/-/compare/yaqc-2020.07.3...yaqc-2020.10.0
 [2020.07.3]: https://gitlab.com/yaq/yaq-python/-/compare/yaqc-2020.07.2...yaqc-2020.07.3
 [2020.07.2]: https://gitlab.com/yaq/yaq-python/-/compare/yaqc-2020.07.1...yaqc-2020.07.2
 [2020.07.0]: https://gitlab.com/yaq/yaqc-python/-/compare/v2020.06.2...v2020.07.0